requirements.txt
dotenv
requests
google-generativeai